+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Light/dark mode toggle.
 
+### Changed
+- API for DNS changed to a show-all approach - table adjusted to suit.
+
 ## [0.2.22] - 2023-01-25
 ### Changed
 - Upgraded dependencies.