+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DNS Inspector is now Domain Tools.
 - API for DNS changed to a show-all approach - table adjusted to suit.
 
+### Fixed
+- Some elements were appearing above the header bar.
+
 ## [0.2.22] - 2023-01-25
 ### Changed
 - Upgraded dependencies.