+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Latest
+### Fixed
+- Fixed the IP inspector not appearing when A records are numerous.
+
+## [0.2.23] - 2023-02-18
 ### Added
 - Light/dark mode toggle.